fractalsql/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
python-dotenv==1.0.1
pandas==2.2.3
openpyxl==3.1.5
filelock==3.16.1
